Updated the Lock() class to allow for self cleanup when a script is killed/terminated and the orphaned lock file is left behind using a process to uniquely identify whether or not the process is still running.

+++ phpext/blazephp/blazephp/lock.zep
@@ -18,12 +18,29 @@ namespace BlazePHP;
 class Lock extends Struct
 {
 	private lockFileLocation;
+	private isLockOwner = false;
 
 	public function __construct(procName)
 	{
-		let this->lockFileLocation = ABS_VAR . "/lock/" . procName . ".lock";
+		let this->lockFileLocation = ABS_VAR . "/lock/" . preg_replace("/[^a-zA-Z0-9_]/", "", procName) . ".lock";
+		// let this->lockFileLocation = "/tmp/" . preg_replace("/[^a-zA-Z0-9_]/", "_", procName) . ".sock";
 
 		if(file_exists(this->lockFileLocation)) {
+
+			var processLine, processParts, processId, processSignature;
+			let processLine  = file_get_contents(this->lockFileLocation);
+			let processParts = explode("-", processLine);
+			if isset(processParts[0]) {
+				let processId = processParts[0];
+			}
+			if isset(processParts[1]) {
+				let processSignature = processParts[1];
+			}
+			if this->getProcessSignature(processId) != processSignature {
+				unlink(this->lockFileLocation);
+				this->__construct(procName);
+				return;
+			}
 
 			var now, fileTime, diff, fileAge;
 
@@ -45,16 +62,43 @@ class Lock extends Struct
 			);
 		}
 		else {
-			var fp;
-			let fp = fopen(this->lockFileLocation, "w");
-			fwrite(fp, this->lockFileLocation);
+			var fp, processId, procSignature;
+			let processId     = getmypid();
+			let fp            = fopen(this->lockFileLocation, "w");
+			let procSignature = this->getProcessSignature(processId);
+			fwrite(fp, processId . "-" . procSignature);
 			fclose(fp);
+			let this->isLockOwner = true;
 		}
+
+		// Remove the lock file if the script is killed or dies without a proper exit
+		// register_shutdown_function("unlink", this->lockFileLocation);
+	}
+
+	public function getProcessSignature(processId)
+	{
+		var psOutput, procSignature, line;
+		exec("ps ahxwwo pid,command | grep " . processId, psOutput);
+		for line in psOutput {
+			// echo line, "\n";
+			if preg_match("/^" . processId . "/", trim(line)) {
+				return md5(line);
+			}
+		}
+		return "no-process-found-for-pid-" . processId;
+	}
+
+	public function destruct()
+	{
+		this->__destruct();
+		exit();
 	}
 
 	public function __destruct()
 	{
-		unlink(this->lockFileLocation);
+		if this->isLockOwner == true {
+			unlink(this->lockFileLocation);
+		}
 	}
 
 	public function fileLocation()
